+++ lib/friendly_id/scoped.rb
@@ -127,8 +127,7 @@ module FriendlyId
       friendly_id_config.scope_columns.each do |column|
         relation = relation.where(column => send(column))
       end
-      primary_key_name = self.class.primary_key
-      relation.where(self.class.arel_table[primary_key_name].not_eq(send(primary_key_name)))
+      relation.where.not(primary_key_values)
     end
     private :scope_for_slug_generator